Move the weak functions of MatterGetAccessPrivilegeForXX to AccessPrivilegeGet.cpp
