+++ src/app/RequiredPrivilege.cpp
@@ -25,23 +25,3 @@ constexpr Access::Privilege RequiredPrivilege::kPrivilegeMapper[];
 
 } // namespace app
 } // namespace chip
-
-int __attribute__((weak)) MatterGetAccessPrivilegeForReadAttribute(chip::ClusterId cluster, chip::AttributeId attribute)
-{
-    return kMatterAccessPrivilegeAdminister;
-}
-
-int __attribute__((weak)) MatterGetAccessPrivilegeForWriteAttribute(chip::ClusterId cluster, chip::AttributeId attribute)
-{
-    return kMatterAccessPrivilegeAdminister;
-}
-
-int __attribute__((weak)) MatterGetAccessPrivilegeForInvokeCommand(chip::ClusterId cluster, chip::CommandId command)
-{
-    return kMatterAccessPrivilegeAdminister;
-}
-
-int __attribute__((weak)) MatterGetAccessPrivilegeForReadEvent(chip::ClusterId cluster, chip::EventId event)
-{
-    return kMatterAccessPrivilegeAdminister;
-}
